refactor (Command classes): update taskList import statements
