+++ src/main/java/tilo/command/AddDeadlineCommand.java
@@ -1,6 +1,6 @@
 package tilo.command;
 
-import tilo.task.TaskList;
+import tilo.storage.TaskList;
 import tilo.ui.Ui;
 import tilo.task.Deadline;
 import tilo.exception.TiloException;
