+++ src/main/java/tilo/command/ExitCommand.java
@@ -1,11 +1,10 @@
 // ExitCommand.java - Command for exiting the application
 package tilo.command;
 
-import tilo.task.TaskList;
+import tilo.storage.TaskList;
 import tilo.ui.Ui;
 
 public class ExitCommand extends Command {
-
     @Override
     public void execute(TaskList taskList, Ui ui) {
         // No action needed - the isExit() method handles the exit logic
